[x] move cargo selected package into its own module
Going to add more Cargo-related functionality in other modules within cargo,
want to avoid bloating up this file too much.
